+++ inc/stm32_control.h
@@ -14,8 +14,8 @@ char X_V;
 char Y_V;
 
 }CVFEED;
-extern unsigned int velspeed ;
-extern unsigned int angspeed ;
+extern float velspeed ;
+extern float angspeed ;
 extern CVFEED cv_res;
 extern unsigned int positionx;
 extern int *stm_Loop();
@@ -23,7 +23,7 @@ extern void cmd_vel_callback(const char * cmd_vel);
 extern void cmd_send(const char cmd_v,int speed);
 extern void cmd_send2(float vspeed,float aspeed);
 
-
+#include "osp_syslog.h"
 #ifdef __cplusplus
 }
 #endif
